Clippy

+++ src/input.rs
@@ -85,7 +85,7 @@ pub fn convert_screen_to_world_position(
     camera_transform: &GlobalTransform,
 ) -> Vec2 {
     // get the size of the window
-    let window_size = Vec2::new(window.width() as f32, window.height() as f32);
+    let window_size = Vec2::new(window.width(), window.height());
 
     // convert screen position [0..resolution] to ndc [-1..1] (gpu coordinates)
     let ndc = (screen_pos / window_size) * 2.0 - Vec2::ONE;
